:recycle: Rename env to config
Closes #228
