+++ test/setup.js
@@ -17,7 +17,7 @@ import 'babel-polyfill';
 import should from 'should';
 import sinon from 'sinon';
 import 'should-sinon';
-import '../src/utils/env';
+import '../src/utils/config';
 
 process.env.NODE_ENV = 'test';
